fix year issue, and enhance display
